glapi: Allow one function name to be in multiple categories.
The gl.spec file doesn't list the same function name in multiple
categories even when it should (such as glGenQueries between core nad
GL_EXT_timer_query), so it hasn't been handled, even though it is a
bug in piglit_dispatch.  With the gl2ext.h parsing for GLES2 our data
source will be giving us instances of this that we need to handle
successfully, so add support for it now.

This also means we'll be able to add annotation of gl.spec for
additional categories for functions if we find we need to.

Reviewed-by: Paul Berry <stereotype441@gmail.com>