piglit: Always request GNU C extensions.
GNU C has a bunch of useful extensions making C a useful language.
Most of those extensions are also available on the BSDs, but MSVC
loses out.  The solution for MSVC is to write the same wrapper we'd
end up writing per test if we were limiting ourselves to non-GNU C.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>