arb_geometry_shader4: Test input overrun.
Test that linking fails if the geometry shader accesses more components than
are available for every input primitive type.

ARB_geometry_shader4 spec 4.3.6:
"It is illegal to index these varying arrays, or in the case of two-
dimensional arrays, the first range of the array, with a negative integral
constant expression or an integral constant expression greater than or
equal to gl_VerticesIn. A link error will occur in these cases."

Signed-off-by: Fabian Bieler <fabianbieler@fastmail.fm>
Reviewed-by: Paul Berry <stereotype441@gmail.com>