arb_geometry_shader4: Test that adjacent vertices are ignored when no geometry shader is active.
Draw the adjacency version of a primitive in red and blend the normal version
of that primitive in green on top of it. Check that the entire framebuffer is
yellow or black.

Test this with and without indexed drawing (glDrawElements and glDrawArrays).

ARB_geometry_shader4 spec 2.6.1:
"If a geometry shader is not active, the "adjacent" vertices are ignored."

v2: Use generic vertex attribute, don't use coordinate transformation, fix
endian issues, split test loop into seperate invocations of test with command
line argument, consolidate indexed and non-indexed drawing tests in one source
file;

Signed-off-by: Fabian Bieler <fabianbieler@fastmail.fm>