Set config.requires_displayed_window in tests that use front buffers.
When running tests with a non-composited window manager, it isn't safe
to access the front buffer until after the window is shown.  This
patch sets config.requires_displayed_window in tests that access front
buffers, to make sure that the piglit framework exposes the window,
even when testing in "-auto" mode.

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>