core.py: Fix sporadic loss of tests in summary
The checking in TestrunResult to determine if a file was proper json or
not is very fragile: it assumes that the last line of the file will be
}. Sometimes this triggers and rebuilds a valid json file, dropping the
last test and resulting in python overhead.

This patch replaces that check with a try/except block. This block
attempts to load the json file with json.load, and on failure attempts
to fix the json.

V2: - replaces json.loads with json.load: this was causing an error but
	  being caught by the except block.
	- Corrects issue created by the try block changing file.tell() so
	  that file.seek would fail in self.__repairFile()
V3: - refactor to remove duplicate code

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>