copyteximage: Enable test vector printfs even when running in -auto mode.
Each invocation of the copyteximage test exercises a large number of
test vectors; a failure in a single test vector causes the entire test
to fail.

This patch modifies the printfs that show which test vector is being
tested, so that they execute regardless of whether the test is being
run automatically or manually (previously, they only executed when the
test was run manually).  This makes it much easier to troubleshoot
failures when looking at the output of an automated piglit run.

Reviewed-by: Eric Anholt <eric@anholt.net>