Verify that vertex shader attributes are assigned in order starting w/0
THIS IS NOT REQUIRED BY ANY VERSION OF THE OpenGL SPECIFICATION!

However, almost every OpenGL implementation happens to behave this way
when there is a single vertex shader compilation unit linked into the
program.  As a result, some programs accidentally rely on this behavior.
If the application was never tested on a implementation that behaves any
other way, there's a reasonable chance it has bugs without its
developers even knowing.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>