Geometry shaders: test gl_ClipDistanceIn (ARB_geometry_shader4)
Since ARB_geometry_shader4 is written against the OpenGL 2.0
specification (which predates gl_ClipDistance), it does not specify
how gl_ClipDistance should be adapted for use in geometry shaders.
However, there is an obvious analogy to the behaviour of gl_TexCoord
(which, like clip distance, is a vertex shader output whose type is a
1D array).  Specifically, the geometry shader should have access to a
2D input array gl_ClipDistanceIn array, where the first index is the
input vertex, and the second index is analogous to the index on
gl_ClipDistance.

This patch adds piglit tests to ensure that gl_ClipDistanceIn works
properly.

Reviewed-by: Fabian Bieler <fabianbieler@fastmail.fm>