Fix clip-distance-explicit-too-large.frag test
The test incorrectly declared gl_ClipDistance as an "out" variable.
This is incorrect--in a fragment shader it should be an "in" variable.

This wasn't causing a piglit failure, since the test was expected to
produce a compile error.  It just means that it wasn't testing what it
was supposed to.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>