CL: Fix other-data-types image tests
These were throwing errors (windows, catalyst, clc 1.2) because image
data types are not allowed to be declared within a kernel, only as
kernel parameters.

Re-structure the test so that we always test event_t, and progressively
check the additional types that are supported based on the declared CLC
version and image support declaration.

Signed-off-by: Aaron Watry <awatry@gmail.com>