piglit-print-commands: Print relative paths for binaries.
Relative executable paths introduce no new requirements.  The tests'
arguments are already relative, so they already imply that the commands
must be run from the top of the piglit directory for printed commands to
work properly.

The benefit of relative executable paths is that it makes it easier to
run the tests on a different directory later on.

Reviewed-by: Brian Paul <brianp@vmware.com>