glsl: Test that anonymous structure declarators are not allowed
Both GLSL ES 1.00 and GLSL 1.20 disallow "anonymous structures."  Each
spec explains this term, and it's not what I had previously thought.  It
is valid to have a struct that has no name (e.g., struct { int x }
foo;).  It is not valid to have a struct in a declarator without a
variable name.  Some flavors of C allow this inside structures.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Cc: Tom Gall <tom.gall@linaro.org>