junit: Move the old summary code into piglit-summary-junit.py
Since junit.py is an external library used in other projects, there is
little interest in replacing it. However, maintaining multiple summary
backends is tedious, so instead the old summary code is moved into
piglit-summary-junit. This is a valid compromise to keep the old code
(and not screw with what works), and hopefully prevent new output from
being based on it.

Reviewed-by: Jose Fonseca <jfonseca@vmware.com>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>