Expect ARB_texture_cube_map_array/compiler/tex_bias-01.vert to fail.
Bias makes no sense in a vertex shader, as there's no implicitly
computed LOD to add a bias to.  According to the GLSL 4.30 spec, "In all
functions below, the bias parameter is optional for fragment shaders.
The bias parameter is not accepted in any other shader."  1.30 has
similar text disallowing bias in vertex shaders.

Reviewed-by: Paul Berry <stereotype441@gmail.com>