shader_runner: Allow "draw arrays" if no [vertex data] section.
For some geometry shader tests, there's no need for any vertex shader
inputs at all--the geometry shader produces all the data itself.  To
allow tests like these, we need to be able to issue a "draw arrays"
command with any number of input vertices, even if there is no [vertex
data] section.

This patch disables shader_runner's "draw arrays" overflow checks in
the event that there is no [vertex data] section, so that the "draw
arrays" command can be used.  This is safe, since the lack of a
[vertex data] section means that no vertex attrib arrays are enabled,
so the shader won't try to access the vertex input data at all.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>