summary.py: Fix duplicate entries bug
The current behavior is that changes fixes and regressions are
calculated by comparing result[i] against result[i++], of course when
there are four or more results provided a test could regress or be fixed
twice. This causes them to be added to the list multiple times, and thus
written into the HTML or text more than once.

This patch corrects this behavior by converting the list to a set before
passing it along to mako or to be printed. Since a set can only hold
unique values this prevents it from being printed more than once.

Reviewed-by: Chris Forbes <chrisf@ijw.co.nz>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>