builtin_functions.py: Replace tabs with spaces
Currently this file uses a mixture of tabs and white spaces. The tabs
are assumed to be 8 spaces. PEP8 is very clear to *never* use mixed
spaces and tabs, which this file currently does, it is also very clear
that 4 spaces, not 8 should be the standard indent for python files.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>