egl_khr_create_context: Build ES tests only when ES 1 and ES 2 are available.
This was, and still is, kind of broken. Previously we were building ES
tests without linking to libGLES* which lead to the tests being skipped.

Now we're requiring ES 2 for tests that can run with ES 1 or ES 2; or
ES 1, ES 2, or Desktop GL. There is potential for build problems if
building with ES 2 but without ES 1 or Desktop GL.

tests/egl/spec/egl_khr_create_context must have CMakeLists.gles2.txt
(as opposed to CMakeLists.gles1.txt) since piglit dispatch isn't
implemented for ES 1, and we need dispatch.

Reviewed-by: Jordan Justen <jordan.l.justen@intel.com>