threadpool.py: Completely rewrite threadpool.py
This patch completely rewrites the threadpool module. This new
implementation is less complicated than the previous version. It also is
not an external project pulled in, but an original implementation.

This tries to be largely API compatible with the previous ThreadPool
implementation. And for piglit's use case it is, however, it does not
implement the full API supported by the previous implementation.

v2: - fix some comments

Reviewed-and-tested-by: Kenneth Graunke <kenneth@whitecape.org>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>