summary: Include traceback in the summary files.
If the Python framework fails to run a program for some reason, it'll
record a 'fail' status.  It also records a 'traceback' key in the
dictionary.

The new summary code missed this when writing the test detail page,
resulting in tests marked as 'fail' with no indication as to why.

Reviewed-by: Dylan Baker <baker.dylan.c@gmail.com>
Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>