texture-integer: fix issue with GL_RGB_INTEGER_EXT swizzle
The expected alpha was set to zero instead of the random alpha value.
The test passed for all random alpha values greater than zero, because
for those values the resulting alpha was zero. Matching the, incorrect,
expected alpha value of zero. The reason the resulting alpha was zero
was because the computed bias became negative, that in turn caused the
resulting alpha to become negative and then rounded to zero.

The test failed when the random alpha value became zero because it was
expecting the, incorrect, alpha value of zero but got the correct value
of one.

Signed-off-by: Marek Olšák <marek.olsak@amd.com>