Change INVALID_VALUE to INVALID_OPERATION in ARB_vertex_array_bgra test.
In the OpenGL 3.3 specification, this changed from INVALID_VALUE to
INVALID_OPERATION.

According to Ian Romanick, the GL specification provides the guidance
that INVALID_VALUE should be generated for "Numeric argument out of
range".  Since FALSE is in range for a boolean, INVALID_VALUE is not
appropriate, and the change to INVALID_OPERATION was meant as a
correction.

See:
http://lists.freedesktop.org/archives/mesa-dev/2013-July/042509.html

Most implementations support 3.3, so they likely follow the new code.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Cc: Ian Romanick <idr@freedesktop.org>
Cc: Fredrik Höglund <fredrik@kde.org>
Cc: Corey Richardson <corey@octayn.net>