arb_uniform_buffer_object: Add UBO from vs-struct-pad tests
The vs-struct-pad and fs-struct-pad tests both fail on Mesa.  The UBOs
from those tests are added here to help narrow down the origin of the
problem.

This test fails on NVIDIA (304.64 on GTX 260), but it passes on AMD
(Catalyst Legacy 12.06 on a Radeon HD 3650).  The NVIDIA driver
generates {0, 4, 8, 12} for the offsets of ubo1.  Mesa generates the
same (incorrect) offsets.

v2: Added spec quotation to justify the {0, 16, 20, 24}
alignments. Suggested by Paul.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Paul Berry <stereotype441@gmail.com>
Tested-by: Kenneth Graunke <kenneth@whitecape.org>