sso: Test all of the ProgramUniform entry points
This is a nearly exhaustive test of all the glProgramUniform functions
added by GL_ARB_separate_shader_objects.

For every entrypoint, a shader using a uniform of the correct type is
created using glCreateShaderProgramv.  The uniform is then set and
queried (using glGetUniform).  The test passes if the correct data is
returned and no GL errors are generated.

The following aspects of these interfaces are not tested by this test:

    - Uniform arrays set by glProgramUniform*v or by
      glProgramUniformMatrix*v with count > 1.

    - Transpose matrices set by glProgramUniformMatrix*v with transpose
      set to GL_TRUE.

v2: Make the double tests use doubles (instead of unsigned integers).
Copy-and-paste error noticed by Paul.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Acked-by: Paul Berry <stereotype441@gmail.com>