ext_image_dma_buf_import: Fix build with mesa < 9.2
The extension EGL_EXT_image_dma_buf_import is a recent extension,
published 2013-02-09, and therefore the eglext.h on many systems do not
define it.  The first Mesa version in which eglext.h defined the
extension was was Mesa 9.2.

To fix the build on such systems, define the extension's enums in
image_common.h.

Tested-by: Ian Romanick <ian.d.romanick@intel.com>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>