builtin_functions.py: tweak integer constants
This fixes some misleading fs-op-div-* failures when the driver does
integer arithmetic with floating point.  When we compute the reciprocal
of some integers we wind up with inexact floating point results.  So
expressions like X/X result in 0.9999 or such.

This patch tweaks the generator code to use some "nicer" integers
so that X/X comes out as expected.

Note the GLSL 1.10 spec, section 4.1.3:
"However, there is no requirement that integers in the language map to
an integer type in hardware. It is not expected that underlying hardware
has full support for a wide range of integer operations."