Optionally capture dmesg changes for each test and report them in a summary
The Radeon driver writes GPU page faults to dmesg and we need to know which
tests caused them.

If there is any change in dmesg during a test run, the test result is changed
as follows:
* pass -> dmesg-warn
* warn -> dmesg-warn
* fail -> dmesg-fail
Dmesg is captured before and after each test and the difference between the two
is stored in the test summary.

The piglit-run.py parameter which enables this behavior is --dmesg. It's also
recommended to use -c0.

v2: - address some of Dylan's remarks, mainly bug fixes
    - fix get_dmesg_diff

Reviewed-by: Dylan Baken <baker.dylan.c@gmail.com>