Run all tests out of a ThreadPool
This patch removes the ConcurrentThreadPool singleton and opts instead
for creating two ThreadPools, one with multiple threads and a second one
that contains only one thread.

The reasoning for this change is two fold, first it gets rid of a rather
questionable use of a singleton, and second it cleans up the Test
derived classes. since they don't have to have a method to add
themselves to the ThreadPool, instead the ThreadPool has a method for
adding tests to itself.

v2: - Only add concurrent tests to the concurrent ThreadPool
    - Correct typos in comments

Reviewed-and-tested-by: Kenneth Graunke <kenneth@whitecape.org>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>