summary.py: Replace buildDictionary function with a much simpler method
This patch produces the exact same output as the previous approach, just
in a much simpler way. It also moves some more of the status information
out of if branches and into the Status class in the status module.

There are a few side effects of using a defaultdict class, namely that
they do not generate a KeyError when one calls for a non existent key.
This means there are a couple of places where try/except blocks are
replaced by if checks.

v2: - fix HTMLIndex._groupResult()
    - don't overwrite the previous status with skip
    - Use a helper function to reduce code duplication

Tested-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>