piglit-util-gl-common: Add Function to clear extensions list.
The default handling of Extension fall backs does not work in specific
cases where the list of supported extensions changes. The list of
extensions for a Core profile may be different from the same
context created using a Compatibility profile.

V2: applied the changes suggested by Chad to improve this function.
    In my previous submission of this I forgot to make note of this.
    This commit changed the new function to:

    piglit_gl_reinitialize_extensions

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>