sso: Combine 20 vertex shaders and 20 fragment shaders in various ways.
Verify that the right shaders are used in the right combinations several
ways.

* The vertex shader has information baked-in that determines the X position
  of the block on the screen.

* The fragment shader has information baked-in that determines how the
  block is colored.  This is combined with data passed from the vertex
  shader.

Since data is passed from the vertex shader to the fragment shader, the
test can use either rendezvous-by-name (default) or rendezvous-by-location
(with --by-location command line parameter).

This test passes in both modes on NVIDIA (304.64 on GTX 260).

v2: Ensure that the window is sized large enough for the tests.  Use a
much simpler shuffling algorithm.  Both suggested by Paul.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Paul Berry <stereotype441@gmail.com>