framework: stop copying dma buffer source padding
Fixes a bug spotted by Chad.

The initial support allowed the stride of the source content in
the dma buffer creation to have some padding but it had to be
less than the padding in the target dma buffer being created.

Moreover the implementation also unnecessarily copied any
padding present in the source to the destination. By copying
only the actual pixel values one not only fixes this but also
allows the source to have greater padding than the destination.

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>
Signed-off-by: Topi Pohjolainen <topi.pohjolainen@intel.com>