summary.py: Treat subtests as groups
This patch causes tests with subtests to be treated as a group, rather
than as a test. This means that the status the test itself stores will
be overwritten by those in the subtest.

There is one oddity about this to be aware of; a test with subtests that
crashes or fails before any of the subtests run will report a fraction
of 0/1 with the appropriate color, even though all of the subtests will
report Not Run.

v2: - Add subtests to the results file as full tests (the internal view
      of the json), without this they will not appear in changes, fixes,
      etc
    - Render the background color of Not Run tests correctly in HTML
    - Apply subtest fractions down the stack
v3: - Don't generate a link for Not Run tests in html page

Tested-by: Ben Widawsky <ben@bwidawsk.net>
Tested-by: Tom Stellard <thomas.stellard@amd.com>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>