query_renderer: Add some common infrastructure for GLX_MESA_query_renderer tests
All of the tests for this extension will require a bunch of function
pointers, etc.  This patch adds a function to get these function
pointers and some macros to wrap calls using them.

v2: Don't use __ in function names because it is reserved by ISO C.
Suggested by Paul.  FWIW, all of the other GLX tests use __ in the
function pointer names.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Paul Berry <stereotype441@gmail.com>