Add new test for calling glGenQueries/glDeleteQueries with an active query.
It is legal to gen and delete a query while another query is active.

It is also legal to delete the currently active query.

Ensure that both of these operations can be performed without any
error, (as well as ensuring that calling glEndQuery on a deleted query
does generate an error).

Reviewed-by: Brian Paul <brianp@vmware.com>
Tested-by: Brian Paul <brianp@vmware.com>