Rename and clarify interface-blocks-vs-gs-array-size-mismatch.shader_test.
The test was correct, but not for the reasons stated in the comment.
The reason a sized VS output interface array should fail to match an
unsized GS input interface array is because geometry shader inputs
require an extra level of array indexing to match corresponding vertex
shader outputs.

This patch renames the test, and updates the comment, to reflect
what's actually being tested.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>