tests/all.py: load glean glsl, frag, and vert tests internally
This patch converts the glean glsl1, fragPrgo1, and vertProg1 tests from
external files that are called by execfile into lists of arguments, and
uses a loop to add the tests.

This removes test files that are not in fact loadable test files, and
lack a TestProfile instance, which means they will not be importable.

V2: - rename glean_glsl1_tests to glean_glsl_tests
    - rename glean_frag_tests to glean_fp_tests
    - rename glean_vert_Tests to glean_vp_tests
    - split removal of blacklist out of this patch

Acked-by: Daniel Vetter <daniel@ffwll.ch>
Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>