core.py: rewrite loadTestProfile to remove execfile()
This patch removes execfile, replacing it with import instead.

This isn't a prefect solution, but it provides feature parity with the
execfile() approach, while being python3 safe.

With this patch a user can require a test profile as tests/<your
profile>.* since, the code actually drops the '.*' , so scripts that
wrap piglit and use (ie) quick.tests will continue to work. It also
works without a '.', so the following examples will work after this
patch:
    tests/quick.tests
    tests/quick.py
    tests/quick

V2: - add explanation of new acceptable names for tests.

Acked-by: Daniel Vetter <daniel@ffwll.ch>
Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>