glean: Fix occlusion query test: expect Gen/DeleteQueries while active to work
A recent inspection of the latest OpenGL specification revealed that
there is no requirement for errors to be generated if GenQueries or
DeleteQueries are called while a query is active.

Specifically, in Section 4.2 of OpenGL 4.4 (Core Profile) the only
errors specified for GenQueries and DeleteQueries is INVALID_VALUE if
the ID is negative.

Additionally, the same section specifies what to do in case the active
query is deleted:

	If an active query object is deleted its name immediately
	becomes unused,...

Implementing this requires that DeleteQueries be able to succeed when
a query is active.

Reviewed-by: Brian Paul <brianp@vmware.com>