Add a "slow_cc" option to EXT_framebuffer_multisample/alpha-blending.
When fast color clears are added to i965, there is a danger that they
may cause this test to stop reproducing bug 53077.  To avoid that, add
a command-line option which causes the test to use a clear color that
isn't fast color clearable.

As a result of this patch, a new piglit test is added:
"spec/EXT_framebuffer_multisample/alpha-blending slow_cc".  As of Mesa
commit fd05ede, this test is known to fail on Ivy Bridge and Haswell.
These failures are already covered by bug 53077.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>