gbm: add event loop and exit after 1 run
If -auto is not used when the piglit platform is gbm, then the test
will hit an abort call. Additionally, it will not actually call
the test display function.

After this change, the test will run once and exit, similar
to what were to happen if the user had pressed the escape key
with other frameworks.

Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>