Test that all layered framebuffer attachments must be of the same type.
As of commit ec79c05, mesa checks that all layered framebuffer
attachments have the same layer count (which is actually
incorrect--layered framebuffer attachments are allowed to have
different layer counts), but it does not check that all layered
framebuffer attachments have the same type.

Therefore, to exercise this bug, we need a test which uses two layered
attachments with different types but the same layer count.  In this
test, one attachment is a cube map (which by its nature has 6 layers),
and the other is a 2D array texture with 6 layers.

This test is known to fail with Mesa commit ec79c05.

Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>