sample-mask-execution: resolve into fbo always
If the MSAA resolve happens into the default framebuffer, we don't get
the colors we expect.

Adjusts the test to resolve into an FBO first, and then blit the result
to either the test FBO or the default framebuffer.

V2: (Timothy Arceri) GL_FRAMEBUFFER -> GL_DRAW_FRAMEBUFFER, ss_tex -> ss_rb
based on review by Anuj. Tested that this fixes the test on Ivy Bridge.

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>
Reviewed-by: Timothy Arceri <timothy.arceri@collabora.com>
Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>