piglit-resume.py: Fix bug in resume
This patch removes an additional JSONWriter.close_dict() call. This
extra call closed the tests dictionary after writing the tests from the
run into it; meaning when the JSONWriter instance is passed into
TestProfile.run() the tests dictionary is already closed and new tests
are not written into the root dictionary.

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>