piglit_fbo_framework - don't even try for OpenGL ES 1.X
glGenFramebuffers is not supported for OpenGL ES 1.X, so init_gl is
hard coded to always fail for ES1.  Thus, there is no way to use a
fbo_framework w/ ES1.  In this case, just fail faster in
piglit_fbo_framework_create() instead of first initializing a wfl
framework, just to destroy it again in the failure path.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>