minmax: Detect cases where errors are generated or values not written
Any minmax test should fail if the query either generates an error or
doesn't write a value to the destination.  To facilitate this, the
storage location is initialized to a value that is unlikely to be
generated by the GL (0xDEADBEEF) instead of 9999.

This caused some additional failures arb_viewport_array-minmax to be
flagged while the extension was being developed.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>