arb_texture_view: Fill 2D array texture properly
The depth only varies by miplevel for volume textures -- all miplevels
have the same number of layers for an array texture.

Previously we would fail to fill all but the first layers, and then
sample from whatever junk was in that memory. This might have worked by
accident on some hardware.

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>
Reviewed-by: Jon Ashburn <jon@lunarg.com>