Fix GleanTest.globalParams
The problem with the way global params work is that they surprise the
user, since they are applied at object initialization time. So the
following code doesn't do what one expects:
   a = GleanTest("basic")
   GleanTest.globalParams = ['--quick']
   b = GleanTest("basic")
   assert a.command == b.command

This would create an assertion failure with the current code, since
global params are applied at instance initialization. Since these params
are supposed to be global, one would expect them to be applied to all
instances, regardless of when the instance was created and the class
variable set.

This patch achieves that by converting Test to a new-type class,
allowing getters and setters to be created. It then creates a getter and
setter for ExecTest.command, and overwrites the getter in GleanTest to
return ``self.command + self.globalParams''. This allows an individual
gleantest to overwrite the global commands, but means that if a instance
hasn't created self.globalParams that the Gleantest.globalParams is
applied when self.command is called.

Acked-by: Jose Fonseca <jfonseca@vmware.com>
Tested-by: Michel Dänzer <michel.daenzer@amd.com>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>